requirements.txt
notion-client~=2.7.0
icalendar~=6.3.2
python-dotenv~=1.2.1